UIEngine이 실수 길이를 제대로 지원하지 않던 점 개선

+++ UIEngine.h
@@ -14,7 +14,7 @@ typedef enum {
 
 typedef struct {
 	UILengthType Type;
-	int Constant;
+	float Constant;
 
 	float Evaluated;
 } UILength;
